+++ frontend/src/components/mypage/Modal.tsx
@@ -1,0 +1,56 @@
+import React, { useEffect } from "react";
+import { FiX } from "react-icons/fi";
+
+interface ModalProps {
+  isOpen: boolean;
+  onClose: () => void;
+  title: string;
+  children: React.ReactNode;
+}
+
+const Modal: React.FC<ModalProps> = ({ isOpen, onClose, title, children }) => {
+  // ESC 키로 모달 닫기
+  useEffect(() => {
+    const handleEscapeKey = (event: KeyboardEvent) => {
+      if (event.key === "Escape" && isOpen) {
+        onClose();
+      }
+    };
+
+    if (isOpen) {
+      document.addEventListener("keydown", handleEscapeKey);
+    }
+
+    return () => {
+      document.removeEventListener("keydown", handleEscapeKey);
+    };
+  }, [isOpen, onClose]);
+
+  if (!isOpen) return null;
+
+  return (
+    <div className="fixed inset-0 z-50 flex items-center justify-center">
+      {/* 배경 오버레이 */}
+      <div className="absolute inset-0 bg-black bg-opacity-50" onClick={onClose} />
+
+      {/* 모달 컨텐츠 */}
+      <div className="relative bg-white dark:bg-[#292929] rounded-lg shadow-xl max-w-4xl w-full mx-4 max-h-[90vh] overflow-hidden">
+        {/* 모달 헤더 */}
+        <div className="flex items-center justify-between px-4 py-3 sm:px-6 sm:py-3 border-b border-gray-200 dark:border-gray-500">
+          <h2 className="text-lg sm:text-xl font-semibold text-gray-900 dark:text-white">{title}</h2>
+          <button
+            onClick={onClose}
+            className="p-1.5 hover:bg-gray-100 dark:hover:bg-gray-700 rounded-full transition-colors"
+          >
+            <FiX className="w-4 h-4 text-gray-500 dark:text-gray-400" />
+          </button>
+        </div>
+
+        {/* 모달 바디 */}
+        <div className="p-4 sm:px-6 sm:py-3 overflow-y-auto max-h-[calc(90vh-60px)]">{children}</div>
+      </div>
+    </div>
+  );
+};
+
+export default Modal;
